+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 343
+#define BUILD 344